refactor: auth, 경력 도메인 리팩토링 (#130) BACKEND-12
