--- src/common/dto/auth/response/login.reponse.ts
@@ -1,6 +1,0 @@
-export class LoginResponse {
-    data: {
-        accessToken: string;
-        refreshToken: string;
-    };
-}
