+++ src/common/dto/users/request/update.user.with.experience.request.ts
@@ -2,6 +2,7 @@ import { ApiProperty } from '@nestjs/swagger';
 
 import { Type } from 'class-transformer';
 import { ValidateNested } from 'class-validator';
+import { UpdateUserExperienceListRequest } from '../../userExperiences/request/update.userExperience.list.request';
 
 import { UpdateUserRequest } from './update.user.request';
 import { UpdateUserExperienceListRequest } from '../../userExperiences/request/update.userExperience.list.request';
