--- src/core/userExperiences/category/category.category.ts
@@ -1,4 +1,0 @@
-export enum UserExperienceEmployment {
-    INTERN = '인턴',
-    FULL_TIME = '정규직',
-}
